requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.29.0
dj-rest-auth==2.2.5
Django==4.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
drf-spectacular==0.23.1
gunicorn==20.1.0
idna==3.3
importlib-resources==5.9.0
inflection==0.5.1
jsonschema==4.15.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pyrsistent==0.18.1
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.8.1
